// src/angular/angular-meixner/src/swagger-api/model/userResponse.ts
/**
 * The Meixner Project
 * The REST API of the Meixner education project
 *
 * OpenAPI spec version: 1.0
 * Contact: mate.szoke2@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserResponse { 
    id: number;
    role: UserResponse.RoleEnum;
    username: string;
}
export namespace UserResponse {
    export type RoleEnum = 'ADMIN' | 'STUDENT';
    export const RoleEnum = {
        ADMIN: 'ADMIN' as RoleEnum,
        STUDENT: 'STUDENT' as RoleEnum
    };
}